Allowing any polynomial order and choosing between them
Any polynomial order can now be selected, but the implementation outside of the following specific cases is rather slow:
- Polynomial order Q = 1, 2, 3 or 4
- Target dimension (or subset dimension) d= 1
This operation is only required once per expectation, so it shouldn't affect the overall time too much.

The cross-validation method has been updated so that if the polyorder input is Inf, then the method will keep increasing the polynomial order until the MSE in the holdout set starts to decrease.